moved checksum method, updated tests accordingly
